+++ __tests__/index.ts
@@ -1,3 +1,3 @@
-export * as cloudrun from './cloudrun/index'
-export * as cloudrunv2 from './cloudrunv2/index'
-export * as compute from './compute/index'
+export * as cloudrun from './cloudrun'
+export * as cloudrunv2 from './cloudrunv2'
+export * as compute from './compute'