+++ src/cloudrun/index.ts
@@ -1,5 +1,7 @@
-import { cloudrunDisallowPublicIngress } from "./cloudrunDisallowPublicIngress";
+import { disallowPublicIngress } from "./disallowPublicIngress";
+import { disallowEnvsSecrets } from "./disallowEnvsSecrets";
 
 export const cloudrunPolicies = [
-	cloudrunDisallowPublicIngress,
+	disallowEnvsSecrets,
+	disallowPublicIngress,
 ];