+++ CHANGELOG.md
@@ -12,6 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CodeQL analysis (#4).
 - dependabot npm scan (#4).
 - `cloudrunv2-disallow-public-ingress` (#5).
+- disallow environment variables from Secret Manager - `cloudrun.Service`, `cloudrunv2.Service` and `cloudrunv2.Job` (#6).
+
+### Changed
+
+- Update test and code file structure (#6).
+
+### Fixed
+
+- codeql event on push in main branch (#6).
 
 ## [1.1.0] - 2023-12-25